In progress: behaviours implementation

+++ src/main/java/dev/sergevas/iot/robotics/kipisopych/bot/adapter/in/web/face/FaceDynamics.java
@@ -1,4 +1,4 @@
 package dev.sergevas.iot.robotics.kipisopych.bot.adapter.in.web.face;
 
-public record FaceDynamics(int sleep, int time) {
+public record FaceDynamics(long sleep, int time) {
 }
